deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
plotly
Pillow
imblearn
scikit-learn==1.0.1
feature-engine==1.5.2
tensorflow-cpu